requirements.txt
langchain
beautifulsoup4
requests
faiss-cpu
chromadb
python-dotenv
tiktoken
sentence-transformers